+++ tests/test_sql_book.py
@@ -460,7 +460,7 @@ class TestZeroRead:
         mysession = Session()
         sheet = SQLTableReader(mysession, Pyexcel)
         data = sheet.to_array()
-        content = [[]]
+        content = [[], []]
         # 'pyexcel' here is the table name
         assert list(data) == content
         mysession.close()